requirements.txt
flask
numpy
scikit-learn
gunicorn
python-dotenv
google-generativeai